TraceMarker: do not make assertions about fd number
This check seems to be testing that an error occurred in opening the
file when the test ran, which is an odd thing to assert. It is causing
flakes when run in the Android test infra, and doesn't seem to provide
any meaningful benefit, so let's remove it.

(This is a speculative fix, as I have been unable to reproduce the flake
locally.)

BUG=b:456347376
TEST=$ atest --rerun-until-failure=100 \
           libchrome-gestures_test:TraceMarkerTest

Change-Id: I6c3c736ef244a804270c66317852865f23dc1250
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/7120918
Tested-by: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Code-Coverage: Sean O'Brien <seobrien@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Kenneth Albanowski <kenalba@google.com>