Unit tests: allow checking multiple gestures from one SyncInterpret
To test my fix for b/433623598, I'll need to be able to assert that a
single call to SyncInterpret produces multiple gestures. To do this,
introduce a new TestInterpreterWrapper::SyncInterpretMulti method that
can return multiple gestures, and modify ::SyncInterpret to use the same
mechanism but expect only one gesture to be produced.

This additional expectation exposes problems with MouseInterpreterTest.
Previously, #SimpleTest was only getting the second wheel gesture
produced, which was the horizontal one. This means that `EXPECT_GT(1,
gs->details.wheel.dy)` was passing because dy on the horizontal wheel
gesture was 0. `EXPECT_LT(-1, gs->details.wheel.dx)` was passing because
-1 < 41.5 (the value being produced for the horizontal scroll), but I
suspect there was also a sign mistake here.

MouseInterpreterTest#EmulateScrollWheelTest's [0] assertions on its last
SyncInterpret call were passing because it was only getting the last of
four gestures produced, which happened to be the one with the button
changes it was expecting. It was ignoring two mouse wheel gestures
resulting from the wheel fields being set in the last HardwareState,
which I assume was a mistake and have corrected. However, it was also
ignoring an additional button change gesture reporting the middle button
going down and then up at the same time. This appears to be an actual
bug in the implementation, which I cannot fix as I do not have a
ChromeOS build set up. In addition, the test is enforcing that mouse
wheel emulation produces an inconsistent stream of gestures —
specifically, only the left button being reported as going down, but
both left and right then being reported as lifted. For these reasons,
disable that test and add a TODO to fix the implementation.

[0]: Added in I62d993a998997e78f45a3745160a3fffc94df707, crrev.com/c/4873791

BUG=b:433623598,b:469726780
TEST=$ atest --host libchrome-gestures_test_host

Change-Id: Ibb02e173b8f60199d428ff9e235f5ce85c51261f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/7269128
Code-Coverage: Henry Barnor <hbarnor@chromium.org>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>