ScrollManager: Return std::optional<Gesture> from FillResultFling
This is clearer than using an output parameter which the method may or
may not touch. It also allows us to mark the method with [[nodiscard]],
which helps ensure that any Gesture returned is actually used in some
way. This is useful because the method has changed some state on the
assumption that the fling it returns is used.

BUG=b:245989146
TEST=run unit tests: $ atest --host libchrome-gestures_test_host

Change-Id: Id512faef67166834075864dc81b9616d71b727df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/7274002
Code-Coverage: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>