ImmediateInterpreter: split up DiagonalSnapTest
This was really testing four different behaviours, and even creating a
new interpreter between each one, all in a for loop with if statements.
It's shorter and easier to understand if split up with a fixture and a
simple utility method.

BUG=b:245989146
TEST=run unit tests:
    $ atest --host -c src/immediate_interpreter_unittest.cc

Change-Id: Ia08ae30aea6ce0044cbd2f3acac135dbd19f03a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/7277465
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>
Code-Coverage: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Henry Barnor <hbarnor@chromium.org>